[ConstraintSystem] NFC: Move `OverloadChoice.h` to `include/swift/Sema`
