+++ lib/Sema/Constraint.h
@@ -19,11 +19,11 @@
 #define SWIFT_SEMA_CONSTRAINT_H
 
 #include "CSFix.h"
-#include "OverloadChoice.h"
 #include "swift/AST/FunctionRefKind.h"
 #include "swift/AST/Identifier.h"
 #include "swift/AST/Type.h"
 #include "swift/Basic/Debug.h"
+#include "swift/Sema/OverloadChoice.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/ilist.h"
 #include "llvm/ADT/ilist_node.h"
